airflow/scripts/requirements.txt
requests
python-dotenv
kafka-python
pandas
pyspark==3.3.2
elasticsearch
